Add a backend javascript code structure
The current stack for the backend javascript includes using Backbone (since
it is shipped with WordPress), Knockout.js, and KnockBack, which combines
the power of Knockout's data-binding with Backbone's modeling capabilities.

We also use Browserify to combine all modules together into one big bundle with
the following command:

  browserify app.js > bundle.js

This is very much a Work In Progress as I'm not yet settled on this stack
and often want to quit my job and move to the mountains instead of figuring
out the right solution. But hey - it's a start.

Also, this is attempting to use the WordPress API, which is still very much
alpha software. We may need to use the typical non-restful WordPress AJAX hooks
until the API gets its shit together later on this year.