Add WP API CPT endpoints
This is very much a Work In Progress, as the WP API is alpha software and
may not be released until much later this year with an unstable spec.

We may be better off ditching this entirely for now and going with the
typical, non-restful WP AJAX admin request hooks until WP API gets its shit
together.