Switch to Rivets and Sightglass for data-binding
After much frustration, Rivets seems to be a data-binding framework that
I feel like I understand. We're also moving away from Backbone-relational
because it wasn't working with our browserify implementation, but that could have
been because the bower_components folder wasn't at the same level as the package.json
file, which seems to have cause some oddities in development.

I may switch back to Backbone-relational at some point, but for now, this is a win.